fix: ajouter SoftDeletes et authorization au modèle Entretien et au contrôleur

+++ app/Http/Controllers/EntretienController.php
@@ -11,6 +11,7 @@ class EntretienController extends Controller
 {
     public function index(Candidature $candidature)
     {
+        $this->authorize('viewAny', Entretien::class);
         $entretiens = $candidature->entretiens()->latest()->get();
 
         return view('entretiens.index', compact('candidature', 'entretiens'));
@@ -18,11 +19,13 @@ class EntretienController extends Controller
 
     public function create(Candidature $candidature)
     {
+        $this->authorize('create', Entretien::class);
         return view('entretiens.create', compact('candidature'));
     }
 
     public function store(StoreEntretienRequest $request, Candidature $candidature)
     {
+
         $candidature->entretiens()->create($request->validated());
 
         return redirect()->route('candidatures.entretiens.index', $candidature)
@@ -31,11 +34,13 @@ class EntretienController extends Controller
 
     public function show(Candidature $candidature, Entretien $entretien)
     {
+        $this->authorize('view', $entretien);
         return view('entretiens.show', compact('candidature', 'entretien'));
     }
 
     public function edit(Candidature $candidature, Entretien $entretien)
     {
+        $this->authorize('update', $entretien);
         return view('entretiens.edit', compact('candidature', 'entretien'));
     }
 
@@ -49,6 +54,7 @@ class EntretienController extends Controller
 
     public function archiver(Candidature $candidature, Entretien $entretien)
     {
+        $this->authorize('archiver', $entretien);
         $entretien->archiver();
 
         return redirect()->route('candidatures.entretiens.index', $candidature)
@@ -57,11 +63,13 @@ class EntretienController extends Controller
 
     public function forceDelete(Candidature $candidature, Entretien $entretien)
     {
+        $this->authorize('forceDelete', $entretien);
         $entretien->forceDelete();
     }
 
     public function restore(Candidature $candidature, Entretien $entretien)
     {
+        $this->authorize('restore', $entretien);
         $entretien->restore();
 
         return redirect()->route('candidatures.entretiens.index', $candidature)
@@ -70,6 +78,7 @@ class EntretienController extends Controller
 
     public function destroy(Candidature $candidature, Entretien $entretien)
     {
+        $this->authorize('delete', $entretien);
         $entretien->delete();
 
         return redirect()->route('candidatures.entretiens.index', $candidature)
